NDS-1099: Extract the styles out for readability
